week_2_workflow_orchestration/docker-requirements.txt
pandas==1.5.2
prefect-sqlalchemy==0.2.2
prefect-gcp[cloud_storage]==0.2.3
protobuf==4.21.11
pyarrow==10.0.1
